requirements.txt
# C:/Users/Sam/PycharmProjects/YAKtunerCONVERTED/requirements.txt
streamlit==1.46.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scipy==1.12.0
asteval==0.9.31
gspread==5.12.4
google-auth-oauthlib==1.2.0
google-auth==2.29.0
st-copy-button
google-generativeai
faiss-cpu
PyMuPDF
python-dotenv